feat: adjust admin dashboard

+++ src/lib/actions/roles.action.ts
@@ -2,11 +2,18 @@
 
 import { and, desc, eq, ilike } from 'drizzle-orm';
 
+import { checkAdmin } from '@/lib/auth';
 import { SearchRole } from '@/lib/schema';
 
 import { database, role } from '@/database';
 
 export async function getRoles(search: SearchRole) {
+    const { isAdmin } = await checkAdmin();
+
+    if (!isAdmin) {
+        return [];
+    }
+
     return await database
         .select()
         .from(role)
@@ -15,6 +22,12 @@ export async function getRoles(search: SearchRole) {
 }
 
 export async function editRoleName(roleId: string, name: string) {
+    const { isAdmin } = await checkAdmin();
+
+    if (!isAdmin) {
+        return;
+    }
+
     return await database
         .update(role)
         .set({
@@ -24,10 +37,22 @@ export async function editRoleName(roleId: string, name: string) {
 }
 
 export async function deleteRole(roleId: string) {
+    const { isAdmin } = await checkAdmin();
+
+    if (!isAdmin) {
+        return;
+    }
+
     return await database.delete(role).where(eq(role.id, roleId));
 }
 
 export async function createRole(name: string) {
+    const { isAdmin } = await checkAdmin();
+
+    if (!isAdmin) {
+        return;
+    }
+
     return await database
         .insert(role)
         .values({
